add repository

+++ app/src/main/java/com/arjun/food2fork/model/Recipe.kt
@@ -19,7 +19,7 @@ data class Recipe(
     @Json(name = "recipe_id")
     var recipeId: String?,
     @Json(name = "social_rank")
-    var socialRank: Int?,
+    var socialRank: Float?,
     @Json(name = "source_url")
     var sourceUrl: String?,
     @Json(name = "title")
